Rename file from PasswordGeneratorField to PasswordField
